refactor: extract git branch helpers
